test: Consolidate test modules for mass matrix.
